# cmd/agents-server/internal/docs/swagger.yaml
components:
  schemas:
...
    handler.ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/protocol.APIError'
      type: object
...
    handler.SettingView:
      properties:
        key:
          type: string
        unknown:
          description: |-
            Unknown marks a row the registry does not define (a removed feature's).
            Reads list it so it can be seen and deleted; writes refuse it.
          type: boolean
        value:
          type: string
      type: object
...
    handler.setSettingReq:
      properties:
        value:
          type: string
      type: object
...
    protocol.APIError:
      properties:
        code:
          type: string
        message:
          type: string
      type: object
...
    settings.Def:
      properties:
        default:
          description: |-
            Default is the value that applies when the setting is unset. Empty means
            the feature is off — never "the zero value happens to be right".
          type: string
        description:
          type: string
        group:
          type: string
        key:
          type: string
        kind:
          $ref: '#/components/schemas/settings.Kind'
        label:
          type: string
        max:
          type: integer
        min:
          description: Min and Max bound a KindInt value. A zero Max means unbounded.
          type: integer
        placeholder:
          type: string
      type: object
    settings.Kind:
      enum:
      - string
      - text
      - secret
      - int
      - bool
      type: string
      x-enum-comments:
        KindSecret: masked on read, sentinel-resolved on write
        KindString: single-line free text
        KindText: multi-line free text
      x-enum-varnames:
      - KindString
      - KindText
      - KindSecret
      - KindInt
      - KindBool
...
  securitySchemes:
    BearerAuth:
      description: Value is "Bearer <token>" — the token printed at server startup
        (or set via --token).
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: 'REST API of agents-server, the web server over the agents-go SDK:
    sessions, runs, agents, MCP servers, memories, settings, skills, providers, workflows,
    triggers, guardrails, sandboxes and projects. A run starts over REST (POST /sessions/{id}/runs)
    or the WebSocket protocol (docs/reference/protocol.md) and streams over either;
    this spec covers the HTTP surface.'
  title: agents-server API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /setting-defs:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/settings.Def'
                type: array
          description: OK
      security:
      - BearerAuth: []
      summary: List setting definitions
      tags:
      - settings
  /settings:
    get:
      description: Every stored key/value. Secrets are masked; a key the registry
        no longer defines is flagged `unknown` with its value masked too (whether
        it was a secret is unknowable), so it can be deleted. The definitions themselves
        are at /setting-defs.
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/handler.SettingView'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: List settings
      tags:
      - settings
  /settings/{key}:
    delete:
      parameters:
      - description: Setting key
        in: path
        name: key
        required: true
        schema:
          type: string
      responses:
        "204":
          description: deleted
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Delete setting
      tags:
      - settings
    get:
      parameters:
      - description: Setting key
        in: path
        name: key
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.SettingView'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get setting
      tags:
      - settings
    put:
      description: The key must be one the registry defines (see /setting-defs) and
        the value must suit its kind; either failure is a 400. An empty value returns
        the setting to its default.
      parameters:
      - description: Setting key
        in: path
        name: key
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.setSettingReq'
                description: Value to store
                summary: setting
        description: Value to store
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.SettingView'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Set setting
      tags:
      - settings
...
servers:
- url: /api/v1